W-0098: a world space id is 128 bits, decided before the bytes froze
Review caught a conflict that an exact binary contract would have cemented. `WorldSpaceId`
was 32 bits, and `W-0104` mints a world space per live dungeon instance -- at runtime, from
whichever region is asked. A 32-bit id needs a central allocator or hand-managed ranges to
stay unique, and the failure mode of getting that wrong is two parties sharing a space and
standing on each other's tiles. "UUID-like" in W-0104 means wide enough to mint without
coordinating, so the field is now `u128` and 16 little-endian bytes on the wire. The position
record grows from 17 bytes to 29.

`RegionId` stays 32 bits deliberately: regions come from the topology manifest, are allocated
there and are never reused, so they need no independent minting.

Three fixture cases exist only to prove the width is real -- 2^32, 2^64+1 and u128::MAX as
space ids, which a narrower field would truncate to the same value. Both languages encode and
decode all three, and a test asserts that three spaces differing only above bit 32 produce
three different byte strings.

Regenerating the fixture rather than patching it, after a patch-on-patch attempt wrote the
literal characters `\\x00` into the rejection lines instead of a zero byte. The generator is a
script now, so the next width change is one edit rather than eleven.

Also recorded in W-0093: `StubHandlersTest` "double_click too far away sends distance error"
failed once under full-suite load with an empty message list, passing in isolation and on
rerun. That is the third order- or timing-dependent test observed in this session, which is
itself the problem -- a red gate is ambiguous when three known flakes can produce it, so the
pattern is now recorded alongside the individual tests.

192 ao-core tests, 833 server tests green on a clean run, ./build.sh check 0.

+++ client-rs/crates/ao-core/src/identity.rs
@@ -333,7 +333,7 @@ impl TopologyAnswer {
 mod tests {
     use super::*;
     use crate::position::LocalPosition;
-    use crate::topology::{Geometry, CORE_X, CORE_Y, PITCH_X};
+    use crate::topology::{Geometry, CORE_X, PITCH_X};
     use std::collections::BTreeMap;
 
     fn owner() -> Ownership {
@@ -555,8 +555,8 @@ mod tests {
         let region = RegionId(42);
 
         assert_eq!(entity.0, instance.0);
-        assert_eq!(template.0, space.0);
-        assert_eq!(region.0, space.0);
+        assert_eq!(template.0 as u128, space.0);
+        assert_eq!(region.0 as u128, space.0);
 
         // Ownership carries a region and an epoch, and neither is the entity's identity: the
         // entity outlives every owner it has.
